[C#] Optimize union getter
When using new() on a generic struct type T in C#, the compiler
generates a call to Activator.CreateInstance<T>() instead of simply
default-initializing the struct, which is slower and creates garbage.
This is needed for compatibility because futures versions of the
language may add parameter-less construstors to structs (for more
details see https://github.com/dotnet/roslyn/issues/8677).
In the case of FlatBuffer tables we can always default-initialize the
struct since we will call __init afterwards.